Refactored default init and base manipulation
